+++ @pageobject/class/src/findElement.ts
@@ -6,26 +6,29 @@ export interface Adapter<TElement> {
   getCurrentUrl(): Promise<string>;
 }
 
-export type Predicate<TElement> = (
+export type Predicate<TElement, TAdapter extends Adapter<TElement>> = (
+  adapter: TAdapter,
   element: TElement,
   index: number,
   elements: TElement[]
 ) => Promise<boolean>;
 
-export interface PathSegment<TElement> {
+export interface PathSegment<TElement, TAdapter extends Adapter<TElement>> {
   readonly selector: string;
   readonly unique: boolean;
-  readonly predicate?: Predicate<TElement>;
+  readonly predicate?: Predicate<TElement, TAdapter>;
 }
 
-class ElementFinder<TElement> {
-  private readonly adapter: Adapter<TElement>;
+class ElementFinder<TElement, TAdapter extends Adapter<TElement>> {
+  private readonly adapter: TAdapter;
 
-  public constructor(adapter: Adapter<TElement>) {
+  public constructor(adapter: TAdapter) {
     this.adapter = adapter;
   }
 
-  public async findElement(path: PathSegment<TElement>[]): Promise<TElement> {
+  public async findElement(
+    path: PathSegment<TElement, TAdapter>[]
+  ): Promise<TElement> {
     if (path.length === 0) {
       throw new Error('No path segments found');
     }
@@ -45,8 +48,8 @@ class ElementFinder<TElement> {
   }
 
   public async findElements(
-    {selector, predicate}: PathSegment<TElement>,
-    remainingPath: PathSegment<TElement>[]
+    {selector, predicate}: PathSegment<TElement, TAdapter>,
+    remainingPath: PathSegment<TElement, TAdapter>[]
   ): Promise<TElement[]> {
     const parent =
       remainingPath.length > 0
@@ -59,20 +62,25 @@ class ElementFinder<TElement> {
       return elements;
     }
 
-    const results = await Promise.all(elements.map(predicate));
+    const results = await Promise.all(
+      elements.map(async (element, index) =>
+        predicate(this.adapter, element, index, elements)
+      )
+    );
 
     return elements.filter((element, index) => results[index]);
   }
 }
 
-export async function findElement<TElement>(
-  path: PathSegment<TElement>[],
-  adapter: Adapter<TElement>
+export async function findElement<TElement, TAdapter extends Adapter<TElement>>(
+  path: PathSegment<TElement, TAdapter>[],
+  adapter: TAdapter
 ): Promise<TElement> {
   const maybeTimeout = process.env.ELEMENT_SEARCH_TIMEOUT;
 
   return retryOnError(
-    async () => new ElementFinder(adapter).findElement(path),
+    async () =>
+      new ElementFinder<TElement, TAdapter>(adapter).findElement(path),
     1000 / 25 /* 40 ms */,
     maybeTimeout ? parseInt(maybeTimeout, 10) : 5000
   );