+++ @pageobject/selenium-adapter/src/__tests__/predicates.test.ts
@@ -5,6 +5,7 @@ import 'chromedriver';
 
 import {Predicate} from '@pageobject/class';
 import {join} from 'path';
+import {WebElement} from 'selenium-webdriver';
 import {SeleniumAdapter, predicates} from '..';
 import {IndexPage} from '../__fixtures__/IndexPage';
 
@@ -23,7 +24,7 @@ afterEach(async () => {
 });
 
 describe('predicates', () => {
-  let predicate: Predicate<any>;
+  let predicate: Predicate<WebElement, SeleniumAdapter>;
 
   describe('atIndex(n)', () => {
     beforeEach(() => {
@@ -31,12 +32,12 @@ describe('predicates', () => {
     });
 
     it('should return a predicate that returns true', async () => {
-      expect(await predicate(undefined, 1, [])).toBe(true);
+      expect(await predicate(adapter, {} as any, 1, [])).toBe(true);
     });
 
     it('should return a predicate that returns false', async () => {
-      expect(await predicate(undefined, 0, [])).toBe(false);
-      expect(await predicate(undefined, 2, [])).toBe(false);
+      expect(await predicate(adapter, {} as any, 0, [])).toBe(false);
+      expect(await predicate(adapter, {} as any, 2, [])).toBe(false);
     });
   });
 
@@ -48,15 +49,15 @@ describe('predicates', () => {
     it('should return a predicate that returns true', async () => {
       const element = (await adapter.findElements('#foo'))[0];
 
-      expect(await predicate(element, 0, [])).toBe(true);
-      expect(await predicate(element, 1, [])).toBe(true);
+      expect(await predicate(adapter, element, 0, [])).toBe(true);
+      expect(await predicate(adapter, element, 1, [])).toBe(true);
     });
 
     it('should return a predicate that returns false', async () => {
       const element = (await adapter.findElements('#bar'))[0];
 
-      expect(await predicate(element, 0, [])).toBe(false);
-      expect(await predicate(element, 1, [])).toBe(false);
+      expect(await predicate(adapter, element, 0, [])).toBe(false);
+      expect(await predicate(adapter, element, 1, [])).toBe(false);
     });
   });
 });